Made cleaned up version of main6.cpp in src/passthroughTest
Made a new project folder with a cleaned up version of main6.cpp, which I believe is the most recent version of the non-arduino code we had working.

Also contains a project status text file.

Code should be functionally identical, just easier to read and work with, still needs a lot of work.